Rename config files to be consistent
