+++ src/configs/cmapControlConfigs.ts
