+++ src/configs/mapConfigs.ts
