+++ src/hooks/useLayerRegistry.ts
@@ -2,7 +2,7 @@ import { useRef, useCallback, useEffect } from 'react';
 import { TileGrid } from 'ol/tilegrid';
 import { Tile as TileLayer } from 'ol/layer';
 import { XYZ } from 'ol/source';
-import { EXTERNAL_BASELAYERS } from '../configs/mapSettings';
+import { EXTERNAL_BASELAYERS } from '../configs/mapConfigs';
 import { getBaselayerResolutions } from '../utils/layerUtils';
 import { ExternalBaselayer, InternalBaselayer } from '../types/layers';
 import { assertInternalBaselayer } from '../reducers/baselayersReducer';
